+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 2.3.0-beta3
 - Exception statistics improvements and other minor bug fixes. [Full list.] (https://github.com/Microsoft/ApplicationInsights-dotnet-server/milestone/19?closed=1)
+- Cross Components Correlation ID changed from SHA(instrumentation key) to Application ID retrieved from http endpoint `api/profiles/{ikey}/appId`.
 
 ## Version 2.3.0-beta2
 - Automatic collection of first chance exceptions statistics. Use a query like this in Application Analytics to query for this statistics: